+++ frontend/src/app/main/refs.cljs
@@ -636,3 +636,6 @@
 
 (def persistence-state
   (l/derived (comp :status :persistence) st/state))
+
+(def restore
+  (l/derived :restore st/state))